+++ packages/devtools/CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 0.1.14-dev.2
+* Fix issues with async trace event rendering in Timeline.
+* Add timing and id information in Timeline event summary.
+* Improve hint text on connect screen.
 * Update package:vm_service dependency to ^2.2.0.
-
-## 0.1.14-dev.1
-* Fix issues with async trace event rendering.
 
 ## 0.1.13 - 2019-12-10
 * Fix crash opening macOS desktop apps in DevTools.